#2833 - Enhance (external) editors API
- Factor out abstract base class used by ExternalAnnotationEditor and RecogitoHtmlAnnotationEditor
